+++ lib/devicesupport/deviceSupports.go
