+++ lib/devicesupport/deviceSupportsList.go
